PkDeletionStrategy to own file
